refactor some pkg code to minting extensions
