+++ pkg/client/explore.go
@@ -5,14 +5,12 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"strconv"
 
 	"github.com/spf13/cobra"
 
 	"github.com/threefoldtech/rivine/pkg/api"
 	"github.com/threefoldtech/rivine/pkg/cli"
 	"github.com/threefoldtech/rivine/pkg/encoding/siabin"
-	types "github.com/threefoldtech/rivine/types"
 )
 
 func createExploreCmd(client *CommandLineClient) *cobra.Command {
@@ -37,17 +35,8 @@ func createExploreCmd(client *CommandLineClient) *cobra.Command {
 			Long:  "Explore an item on the blockchain, using its hash or ID.",
 			Run:   Wrap(exploreCmd.hashCmd),
 		}
-		getMintConditionCmd = &cobra.Command{
-			Use:   "mintcondition [height]",
-			Short: "Get the active mint condition",
-			Long: `Get the active mint condition,
-either the one active for the current block height,
-or the one for the given block height.
-`,
-			Run: exploreCmd.getMintCondition,
-		}
 	)
-	rootCmd.AddCommand(blockCmd, hashCmd, getMintConditionCmd)
+	rootCmd.AddCommand(blockCmd, hashCmd)
 
 	// create flags
 	blockCmd.Flags().Var(
@@ -63,10 +52,6 @@ or the one for the given block height.
 		&exploreCmd.hashCfg.MinHeight, "min-height", 0,
 		"when looking up the transactions linked to an unlockhash, only show transactions since a given height")
 
-	getMintConditionCmd.Flags().Var(
-		cli.NewEncodingTypeFlag(0, &exploreCmd.getMintConditionCfg.EncodingType, 0), "encoding",
-		cli.EncodingTypeFlagDescription(0))
-
 	// return root command
 	return rootCmd
 }
@@ -80,9 +65,6 @@ type exploreCmd struct {
 	hashCfg struct {
 		EncodingType cli.EncodingType
 		MinHeight    uint64
-	}
-	getMintConditionCfg struct {
-		EncodingType cli.EncodingType
 	}
 }
 
@@ -141,60 +123,5 @@ func (explorerSubCmds *exploreCmd) hashCmd(hash string) {
 		e := json.NewEncoder(os.Stdout)
 		e.SetIndent("", "  ")
 		e.Encode(resp)
-	}
-}
-
-func (explorerSubCmds *exploreCmd) getMintCondition(cmd *cobra.Command, args []string) {
-	txDBReader := NewTransactionDBExplorerClient(explorerSubCmds.cli)
-
-	var (
-		mintCondition types.UnlockConditionProxy
-		err           error
-	)
-
-	switch len(args) {
-	case 0:
-		// get active mint condition for the latest block height
-		mintCondition, err = txDBReader.GetActiveMintCondition()
-		if err != nil {
-			cli.DieWithError("failed to get the active mint condition", err)
-		}
-
-	case 1:
-		// get active mint condition for a given block height
-		height, err := strconv.ParseUint(args[0], 10, 64)
-		if err != nil {
-			cmd.UsageFunc()
-			cli.DieWithError("invalid block height given", err)
-		}
-		mintCondition, err = txDBReader.GetMintConditionAt(types.BlockHeight(height))
-		if err != nil {
-			cli.DieWithError("failed to get the mint condition at the given block height", err)
-		}
-
-	default:
-		cmd.UsageFunc()
-		cli.Die("Invalid amount of arguments. One optional pos argument can be given, a valid block height.")
-	}
-
-	// encode depending on the encoding flag
-	var encode func(interface{}) error
-	switch explorerSubCmds.getMintConditionCfg.EncodingType {
-	case cli.EncodingTypeHuman:
-		e := json.NewEncoder(os.Stdout)
-		e.SetIndent("", "  ")
-		encode = e.Encode
-	case cli.EncodingTypeJSON:
-		encode = json.NewEncoder(os.Stdout).Encode
-	case cli.EncodingTypeHex:
-		encode = func(v interface{}) error {
-			b := siabin.Marshal(v)
-			fmt.Println(hex.EncodeToString(b))
-			return nil
-		}
-	}
-	err = encode(mintCondition)
-	if err != nil {
-		cli.DieWithError("failed to encode mint condition", err)
 	}
 }
